requirements.txt
pandas
numpy
ccxt
python-dotenv
flask
aiohttp
scikit-learn
TA-Lib
joblib
optuna
textblob
tweepy
feedparser
scipy